refactor(onboard): extract sandbox intent types
