+++ src/lib/onboard/sandbox-create-plan.ts
@@ -9,7 +9,21 @@ import type { InitialSandboxPolicy } from "./initial-policy";
 import type { MessagingTokenDef } from "./messaging-prep";
 import type { MessagingChannel } from "./messaging-state";
 import { resolveQrSelectedChannels } from "./messaging-state";
+import type {
+  MaterializeSandboxCreatePlanInput,
+  ResolveSandboxCreateIntentInput,
+  SandboxCreateIntent,
+  SandboxCreateMessagingProviderRequest,
+} from "./sandbox-create-intent-types";
 import { buildSandboxGpuCreateArgs, type SandboxGpuCreateConfig } from "./sandbox-gpu-create";
+
+export type {
+  MaterializeSandboxCreatePlanInput,
+  ResolveSandboxCreateIntentInput,
+  SandboxCreateIntent,
+  SandboxCreateMessagingProviderRequest,
+  SandboxCreatePolicyRequest,
+} from "./sandbox-create-intent-types";
 
 // Known canonical policy tier names. Kept inline so the create-time path
 // validates the env value without pulling `../policy/tiers` (which transitively
@@ -79,84 +93,6 @@ export type SandboxCreatePlan = {
   messagingProviders: string[];
   useDockerGpuPatch: boolean;
   sandboxGpuLogMessage: string | null;
-};
-
-export type SandboxCreateMessagingProviderRequest = {
-  readonly name: string;
-  readonly envKey: string;
-  readonly providerType?: string;
-  readonly credentialConfigured: boolean;
-  readonly channel: string | null;
-};
-
-export type SandboxCreatePolicyRequest = {
-  readonly basePolicyPath: string;
-  readonly activeMessagingChannels: readonly string[];
-  readonly options: {
-    readonly directGpu: boolean;
-    readonly dockerGpuPatch: boolean;
-    readonly additionalPresets: readonly string[];
-    readonly agentName?: string | null;
-    readonly policyTier: string | null;
-  };
-};
-
-/**
- * Serializable intent for the create-time sandbox contributions. When built
- * through {@link prepareSandboxCreatePlan}, messaging credential values are
- * represented only by their logical environment-key bindings and presence.
- *
- * This is deliberately separate from {@link SandboxCreatePlan}, which is an
- * execution artifact containing temporary paths and cleanup callbacks. Its
- * serializable shape is for internal inspection and testing only; it is not a
- * persistence, machine-event, or public API contract.
- */
-export type SandboxCreateIntent = {
-  readonly sandboxName: string;
-  readonly activeMessagingChannels: readonly string[];
-  readonly messagingProviderRequests: readonly SandboxCreateMessagingProviderRequest[];
-  readonly reusableMessagingProviders: readonly string[];
-  readonly extraProviders: readonly string[];
-  readonly hermesToolGateways: readonly string[];
-  readonly policy: SandboxCreatePolicyRequest;
-  readonly gpuCreateArgs: readonly string[];
-  readonly useDockerGpuPatch: boolean;
-  readonly sandboxGpuLogMessage: string | null;
-  readonly disabledChannelNames: readonly string[];
-};
-
-export type ResolveSandboxCreateIntentInput = {
-  basePolicyPath: string;
-  sandboxName: string;
-  channels: MessagingChannel[];
-  enabledChannels: string[] | null;
-  disabledChannelNames: ReadonlySet<string>;
-  messagingProviderRequests: readonly SandboxCreateMessagingProviderRequest[];
-  primaryMessagingCredentialEnvKeys: readonly string[];
-  reusableMessagingChannels: readonly string[];
-  reusableMessagingProviders: readonly string[];
-  extraProviders?: readonly string[];
-  hermesToolGateways: readonly string[];
-  sandboxGpuConfig: SandboxGpuCreateConfig;
-  gpuCreateArgs: readonly string[];
-  useDockerGpuPatch: boolean;
-  sandboxGpuLogMessage: string | null;
-  agentName?: string | null;
-  policyTier: string | null;
-};
-
-export type MaterializeSandboxCreatePlanInput = {
-  intent: SandboxCreateIntent;
-  buildCtx: string;
-  messagingTokenDefs: MessagingTokenDef[];
-  appendResourceFlags(createArgs: string[]): void;
-  runProviderPreDeleteCleanup(): void;
-  upsertMessagingProviders(
-    tokenDefs: MessagingTokenDef[],
-    options: { replaceExisting: true },
-  ): string[];
-  getHermesToolGatewayProviderName(sandboxName: string): string;
-  prepareInitialSandboxCreatePolicy?: PrepareInitialSandboxCreatePolicy;
 };
 
 function getDockerGpuSandboxCreatePlan(
@@ -355,7 +291,8 @@ export function resolveSandboxCreateIntent({
 function messagingProviderRequestKey(
   request: Pick<SandboxCreateMessagingProviderRequest, "name" | "envKey">,
 ): string {
-  return `${request.name}\u0000${request.envKey}`;
+  // Tuple encoding stays collision-free even if either value contains a separator.
+  return JSON.stringify([request.name, request.envKey]);
 }
 
 function bindMessagingTokenDefs(
@@ -382,7 +319,10 @@ function bindMessagingTokenDefs(
         `Cannot materialize sandbox create intent; credential availability changed for provider '${request.name}'.`,
       );
     }
-    if ((tokenDef.providerType || undefined) !== request.providerType) {
+    // Default providers omit this field; normalize an empty or missing binding
+    // to the intent's `undefined` representation before comparing.
+    const boundProviderType = tokenDef.providerType || undefined;
+    if (boundProviderType !== request.providerType) {
       throw new Error(
         `Cannot materialize sandbox create intent; provider type changed for '${request.name}'.`,
       );
